requirements.txt
# AOU Marche HIS - Python Backend Dependencies
# Version: 2.1.0
# Python 3.11+

# ============================================================================
# CORE FRAMEWORKS
# ============================================================================
fastapi==0.104.1              # Modern web framework for APIs
uvicorn[standard]==0.24.0      # ASGI server for FastAPI
gunicorn==21.2.0               # WSGI server for production
python-multipart==0.0.6        # Handle multipart/form-data

# ============================================================================
# AUTHENTICATION & SECURITY
# ============================================================================
python-jose[cryptography]==3.3.0   # JWT token handling
passlib[bcrypt]==1.7.4             # Password hashing
python-jose==3.3.0                  # JWT encoding/decoding
cryptography==41.0.7                # Encryption utilities
pyjwt==2.8.0                        # JSON Web Token implementation
bcrypt==4.0.1                       # Password hashing

# ============================================================================
# DATABASE & STORAGE
# ============================================================================
sqlalchemy==2.0.23              # SQL ORM
psycopg2-binary==2.9.9          # PostgreSQL driver
alembic==1.12.1                  # Database migrations
redis==5.0.1                     # Redis client for caching/sessions
motor==3.3.2                     # Async MongoDB driver
pymongo==4.6.0                   # MongoDB driver
asyncpg==0.29.0                  # Async PostgreSQL driver

# ============================================================================
# MACHINE LEARNING & AI
# ============================================================================
scikit-learn==1.3.2              # ML algorithms for triage
numpy==1.26.2                     # Numerical computing
pandas==2.1.3                     # Data manipulation
joblib==1.3.2                     # Model serialization
tensorflow==2.14.0                # Deep learning (optional)
torch==2.1.1                      # PyTorch (optional)
transformers==4.35.0              # NLP models for symptom analysis
sentence-transformers==2.2.2      # Text embeddings for medical terms

# ============================================================================
# HEALTHCARE STANDARDS
# ============================================================================
fhir.resources==7.0.0             # FHIR resource models
hl7apy==1.3.4                      # HL7 message parsing
python-fhir==3.4.0                  # FHIR client library

# ============================================================================
# REAL-TIME COMMUNICATIONS
# ============================================================================
websockets==12.0                    # WebSocket protocol
socketio==5.11.1                    # Socket.IO server
aiortc==1.6.0                       # WebRTC for Python
opencv-python==4.8.1.78             # Video processing

# ============================================================================
# NOTIFICATIONS & MESSAGING
# ============================================================================
twilio==8.10.0                      # SMS/WhatsApp notifications
firebase-admin==6.4.0                # Push notifications (FCM)
celery==5.3.4                        # Async task queue
redis==5.0.1                          # Message broker for Celery

# ============================================================================
# GEOGRAPHIC SERVICES
# ============================================================================
geopy==2.4.1                         # Geolocation utilities
folium==0.15.1                        # Map generation for ambulance tracking

# ============================================================================
# REPORTING & DOCUMENTS
# ============================================================================
reportlab==4.0.9                      # PDF generation
jinja2==3.1.2                          # Template engine
weasyprint==60.1                       # HTML to PDF conversion
xhtml2pdf==0.2.12                       # Alternative PDF generation

# ============================================================================
# DATA VALIDATION & SERIALIZATION
# ============================================================================
pydantic==2.5.0                       # Data validation
pydantic-settings==2.1.0               # Settings management
pydantic[email]==2.5.0                 # Email validation

# ============================================================================
# HTTP CLIENT & INTEGRATIONS
# ============================================================================
httpx==0.25.1                          # Async HTTP client
aiohttp==3.9.1                         # Async HTTP server/client
requests==2.31.0                        # HTTP client

# ============================================================================
# LOGGING & MONITORING
# ============================================================================
loguru==0.7.2                           # Logging
prometheus-fastapi-instrumentator==6.1.0 # Prometheus metrics
opentelemetry-api==1.21.0                 # OpenTelemetry tracing
opentelemetry-sdk==1.21.0                  # OpenTelemetry SDK

# ============================================================================
# TESTING
# ============================================================================
pytest==7.4.3                            # Testing framework
pytest-asyncio==0.21.1                    # Async tests
pytest-cov==4.1.0                          # Coverage reporting
httpx==0.25.1                               # HTTP client for testing
factory-boy==3.3.0                           # Test data generation
faker==20.1.0                                 # Fake data generation

# ============================================================================
# DEVELOPMENT TOOLS
# ============================================================================
black==23.11.0                            # Code formatting
isort==5.12.0                              # Import sorting
flake8==6.1.0                               # Linting
mypy==1.7.0                                  # Static type checking
pre-commit==3.5.0                            # Git hooks

# ============================================================================
# UTILITIES
# ============================================================================
python-dotenv==1.0.0                       # Environment variables
click==8.1.7                                 # CLI interface
typer==0.9.0                                  # CLI builder
tqdm==4.66.1                                   # Progress bars
pyyaml==6.0.1                                   # YAML parsing